uber: refactor leaky abstraction chan <-bool paradigm to func()
Fixes #11.

The previous abstraction/paradigm for paging cancellation was
leaky and exposed too much information as well as a made it possible
for a user to screw up the program flow by closing the provided channel
or doing any operations on it.

The fix for this is to return a `func()` or cancelling function
that is invoked by the user whenever they need to cancel any paging.
This way the innards of the package can ensure that closing is only
called once and that the user doesn't have to think about channels.